+++ examples/vite/cloudflare/vite.config.ts
@@ -1,19 +1,11 @@
 import { cloudflare } from "@cloudflare/vite-plugin";
 import { srvkit } from "@srvkit/vite/plugin";
+import { cloudflareWorkersPreset } from "@srvkit/vite/presets";
 import { defineConfig } from "vite";
 
 export default defineConfig({
     plugins: [
         cloudflare(),
-        srvkit({
-            runtime: "workerd",
-            build: {
-                target: "handler",
-                public: {
-                    copy: true,
-                    to: "./client",
-                },
-            },
-        }),
+        srvkit(cloudflareWorkersPreset()),
     ],
 });